next and tailwindcss config update, simplified mui chip to custom
