+++ client/next.config.ts
@@ -23,7 +23,10 @@ const nextConfig: NextConfig = {
         pathname: '/**',
         // pathname: '/a/**'
       },
-    ]
+    ],
+    minimumCacheTTL: 86400, 
+    deviceSizes: [640, 750, 828, 1080, 1200, 1920],
+    imageSizes: [16, 32, 48, 64, 96, 128, 256, 384],
   },
   async rewrites() {
     return [
