+++ client/package.json
@@ -8,7 +8,7 @@
     "start": "next start",
     "lint": "next lint",
     "test:prod": "npm run build && npm run start",
-    "analyze": "ANALYZE=true next build"
+    "analyze": "NODE_ENV=production ANALYZE=true next build"
   },
   "dependencies": {
     "@emotion/react": "^11.13.5",
