+++ client/tailwind.config.ts
@@ -7,35 +7,51 @@ export default {
     "./src/app/**/*.{js,ts,jsx,tsx,mdx}",
   ],
   safelist: [
-    'bg-green-600',
-    'bg-green-700',
-    'bg-green-800',
-    'bg-green-900',
-    'bg-slate-300',
-    'bg-slate-600',
-    'bg-slate-700',
-    'bg-slate-800',
-    'bg-slate-900',
-    'bg-saitRed',
-    'bg-saitDarkRed',
-    'bg-saitLighterBlueOg',
-    'bg-saitLighterBlue',
-    'bg-saitLightBlue',
-    'bg-saitBlue',
-    'bg-saitBlueOg',
-    'bg-saitPurple',
-    'bg-saitDarkPurple',
-    'bg-saitLighterPurple',
-    'bg-saitLightPurple',
-    'bg-saitGray',
-    'ng-saitBlack',
-    'bg-saitWhite',
-    'hover:bg-green-500',
-    'hover:bg-slate-800',
-    'text-green-200',
-    'text-green-700',
-    'hover:text-green-200',
-    // Add any other classes you want to safelist
+    { 
+      pattern: /(bg|text|border|hover:bg)-sait(Red|Blue|Purple|Gray|LightRed|LightBlue|LightPurple|DarkRed|DarkBlue|DarkPurple|LighterRed|LighterBlue|LighterPurple)/ 
+    },
+    // all gradient classes
+    'bg-gradient-to-r',
+    'from-slate-700',
+    'via-saitDarkPurple',
+    'to-saitBlue',
+    // image-related classes
+    'object-cover',
+    'rounded-full',
+    'border-4',
+    'border-white',
+    // dynamic classes
+    { pattern: /^text-(sm|base|lg|xl)/ },
+    { pattern: /^font-(normal|semibold|bold)/ },
+    { pattern: /^rounded-(none|sm|md|lg|full)/ },
+    // 'bg-green-600',
+    // 'bg-green-700',
+    // 'bg-green-800',
+    // 'bg-green-900',
+    // 'bg-slate-300',
+    // 'bg-slate-600',
+    // 'bg-slate-700',
+    // 'bg-slate-800',
+    // 'bg-slate-900',
+    // 'bg-saitRed',
+    // 'bg-saitDarkRed',
+    // 'bg-saitLighterBlueOg',
+    // 'bg-saitLighterBlue',
+    // 'bg-saitLightBlue',
+    // 'bg-saitBlue',
+    // 'bg-saitBlueOg',
+    // 'bg-saitPurple',
+    // 'bg-saitDarkPurple',
+    // 'bg-saitLighterPurple',
+    // 'bg-saitLightPurple',
+    // 'bg-saitGray',
+    // 'ng-saitBlack',
+    // 'bg-saitWhite',
+    // 'hover:bg-green-500',
+    // 'hover:bg-slate-800',
+    // 'text-green-200',
+    // 'text-green-700',
+    // 'hover:text-green-200',
 ],
   theme: {
     extend: {
